doc changes, edge node tags
Added a bit about the potential for busted inventory as a result of passing cluster_node_count in extra vars. Also reverted the playbook logic to assume we are using ints. If passing cluster_node_count, it should be in JSON so that the value does not become a string.

Tagged edge node plays so we can skip them when needed.